dashboard/requirements.txt
streamlit==1.32.2
pandas==2.2.1
requests==2.31.0
psycopg2-binary==2.9.9
numpy==1.26.4
